Rename service-worker.ts to worker-messaging.ts
